[Feat] : 로그인 시간 추가, expiration 데이터형 변환

+++ src/main/java/com/potato/balbambalbam/user/token/controller/ReissueController.java
@@ -1,10 +1,13 @@
 package com.potato.balbambalbam.user.token.controller;
 
 import com.potato.balbambalbam.data.entity.Refresh;
+import com.potato.balbambalbam.data.entity.User;
 import com.potato.balbambalbam.data.repository.RefreshRepository;
-import com.potato.balbambalbam.exception.dto.ExceptionDto;
+import com.potato.balbambalbam.data.repository.UserRepository;
 import com.potato.balbambalbam.exception.ResponseNotFoundException;
 import com.potato.balbambalbam.exception.TokenExpiredException;
+import com.potato.balbambalbam.exception.UserNotFoundException;
+import com.potato.balbambalbam.exception.dto.ExceptionDto;
 import com.potato.balbambalbam.user.token.jwt.JWTUtil;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.swagger.v3.oas.annotations.Operation;
@@ -15,6 +18,8 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletResponse;
+import java.time.LocalDateTime;
+import java.util.Date;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -23,8 +28,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.Date;
-
 @RequiredArgsConstructor
 @RestController
 @Slf4j
@@ -32,6 +35,7 @@ import java.util.Date;
 public class ReissueController {
     private final JWTUtil jwtUtil;
     private final RefreshRepository refreshRepository;
+    private final UserRepository userRepository;
 
     @Operation(summary = "토큰 재발급", description = "만료된 또는 유효한 refresh 토큰을 이용하여 새로운 access 및 refresh 토큰을 재발급한다.")
     @ApiResponses(value = {
@@ -41,7 +45,6 @@ public class ReissueController {
     })
     @PostMapping("/reissue")
     public ResponseEntity<?> reissue(@RequestHeader("refresh") String refresh, HttpServletResponse response) {
-
         if (refresh == null) {
             throw new ResponseNotFoundException("refresh 토큰이 없습니다."); // 404
         }
@@ -57,28 +60,34 @@ public class ReissueController {
         String socialId = jwtUtil.getSocialId(refresh);
         String role = jwtUtil.getRole(refresh);
 
-        String newAccess = jwtUtil.createJwt("access", userId, socialId, role, 7200000L);
-        String newRefresh = jwtUtil.createJwt("refresh", userId, socialId, role, 864000000L);
+        // 사용자가 enabled 상태인지 확인
+        User user = userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException("사용자를 찾을 수 없습니다."));
+        if (!user.getEnabled()) {
+            throw new UserNotFoundException("탈퇴한 회원입니다.");
+        }
 
-        //Refresh 토큰 저장 DB에 기존의 Refresh 토큰 삭제 후 새 Refresh 토큰 저장
+        String newAccess = jwtUtil.createJwt("access", userId, socialId, role, 7200000L);
+        String newRefresh = jwtUtil.createJwt("refresh", userId, socialId, role, 8640000000L); // 100일
+
         refreshRepository.deleteBySocialIdAndUserId(socialId, userId);
-        addRefreshEntity(userId, socialId, newRefresh, 864000000L); // 10일
+        addRefreshEntity(userId, socialId, newRefresh, 8640000000L); // 100일
 
         response.setHeader("access", newAccess);
         response.setHeader("refresh", newRefresh);
 
-        return new ResponseEntity<>("refresh 토근과 access 토큰이 재발급 되었습니다.",HttpStatus.OK);//200
+        return new ResponseEntity<>("refresh 토근과 access 토큰이 재발급 되었습니다.", HttpStatus.OK);//200
     }
 
     private void addRefreshEntity(Long userId, String socialId, String refresh, Long expiredMs) {
-
-        Date date = new Date(System.currentTimeMillis() + expiredMs);
+        LocalDateTime now = LocalDateTime.now();
+        LocalDateTime expirationTime = now.plusSeconds(expiredMs / 1000);
 
         Refresh refreshEntity = new Refresh();
         refreshEntity.setUserId(userId);
         refreshEntity.setSocialId(socialId);
         refreshEntity.setRefresh(refresh);
-        refreshEntity.setExpiration(date.toString());
+        refreshEntity.setLastLoginAt(now);
+        refreshEntity.setExpiration(expirationTime);
 
         refreshRepository.save(refreshEntity);
     }
